+++ server/src/constants.ts
@@ -1,0 +1,7 @@
+export enum CustomMessages {
+	CoGStyleGuide = 'choicescript/cogstyleguide',
+	SelectionWordCountRequest = 'choicescript/selectionwordcount',
+	UpdatedProjectFiles = 'choicescript/projectfiles',
+	UpdatedWordCount = 'choicescript/updatedwordcount',
+	WordCountRequest = 'choicescript/wordcount',
+}